raspberry_pi_interface/src/requirements.txt
absl-py==0.7.0
astor==0.7.1
automationhat==0.0.4
blinker==1.3
blinkt==0.1.1
Cap1xxx==0.1.3
chardet==2.3.0
click==6.6
colorama==0.3.7
cryptography==1.7.1
drumhat==0.1.0
envirophat==0.0.6
ExplorerHAT==0.4.2
Flask==0.12.1
fourletterphat==0.1.0
gast==0.2.2
gpiozero==1.4.1
grpcio==1.18.0
h5py==2.9.0
idna==2.2
itsdangerous==0.24
jedi==0.10.2
Jinja2==2.8
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
keyring==10.1
keyrings.alt==1.3
Markdown==3.0.1
MarkupSafe==0.23
mcpi==0.1.1
microdotphat==0.1.3
mote==0.0.3
motephat==0.0.2
numpy==1.16.2
oauthlib==2.0.1
opencv-python==3.4.4.19
pgzero==1.2
phatbeat==0.1.0
pianohat==0.0.5
picamera==1.13
picraft==1.0
piglow==1.2.4
pigpio==1.38
Pillow==4.0.0
protobuf==3.6.1
pyasn1==0.1.9
pycrypto==2.6.1
pygame==1.9.3
pygobject==3.22.0
pyinotify==0.9.6
PyJWT==1.4.2
pyOpenSSL==16.2.0
pyserial==3.2.1
python-apt==1.1.0b5
pyxdg==0.25
PyYAML==3.13
rainbowhat==0.0.2
requests==2.12.4
requests-oauthlib==0.7.0
RPi.GPIO==0.6.3
RTIMULib==7.2.1
scipy==1.2.1
scrollphat==0.0.7
scrollphathd==1.2.0
SecretStorage==2.3.1
sense-emu==1.0
sense-hat==2.2.0
simplejson==3.10.0
six==1.12.0
skywriter==0.0.7
sn3218==1.2.7
spidev==3.3
tensorboard==1.11.0
tensorflow==1.11.0
termcolor==1.1.0
thonny==2.1.16
touchphat==0.0.1
twython==3.4.0
urllib3==1.19.1
virtualenv==16.4.0
Werkzeug==0.14.1
